requirements.txt
pandas==2.3.3
fastapi==0.128.0
uvicorn==0.40.0
google-genai
requests
python-dotenv
chromadb
sentence-transformers
tenacity
